+++ functions.php
@@ -98,6 +98,7 @@ get_template_part( 'lib/taxonomies' );
 get_template_part( 'lib/meta/meta-boxes-instructions' );
 get_template_part( 'lib/meta/meta-boxes' );
 get_template_part( 'lib/meta/cmb2-validation' );
+get_template_part( 'lib/meta/meta-boxes-seo' );
 get_template_part( 'lib/meta/meta-boxes-post' );
 get_template_part( 'lib/meta/meta-boxes-page' );
 get_template_part( 'lib/meta/meta-boxes-page-about' );
@@ -163,6 +164,7 @@ get_template_part( 'lib/functions-custom' );
 get_template_part( 'lib/functions-filters' );
 get_template_part( 'lib/functions-hooks' );
 get_template_part( 'lib/functions-utility' );
+get_template_part( 'lib/functions-seo' );
 
 // Newsletter migration - auto-runs on deployment
 get_template_part( 'lib/newsletter-migration' );